refactor: player 리스트를 구성하는 기능 리팩터링
